# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users": {
      "post": {
        "description": "Create a new user account with unique user_email and user_password business rules.",
        "operationId": "UsersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "User email already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "422": {
            "description": "Business rule violation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Create user",
        "tags": [
          "users"
        ]
      },
      "get": {
        "description": "List users with pagination, filtering, search, and sorting support.",
        "operationId": "UsersController_findMany",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "หมายเลขหน้าที่ต้องการ",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "จำนวนรายการต่อหน้า",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "field ที่ใช้ sort",
            "schema": {
              "enum": [
                "createdAt",
                "updatedAt",
                "user_email",
                "fullName"
              ],
              "type": "string"
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "description": "ทิศทางการ sort",
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "ค้นหาจาก user_email หรือชื่อ",
            "schema": {
              "example": "somchai",
              "type": "string"
            }
          },
          {
            "name": "user_email",
            "required": false,
            "in": "query",
            "description": "กรองด้วย user_email แบบตรงตัว",
            "schema": {
              "format": "email",
              "example": "user@example.com",
              "type": "string"
            }
          },
          {
            "name": "role",
            "required": false,
            "in": "query",
            "description": "กรอง users ที่มี role นี้",
            "schema": {
              "example": "admin",
              "type": "string"
            }
          },
          {
            "name": "includeDeleted",
            "required": false,
            "in": "query",
            "description": "รวมรายการที่ถูก soft delete แล้วหรือไม่",
            "schema": {
              "example": false,
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Users returned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserListResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "List users",
        "tags": [
          "users"
        ]
      }
    },
    "/users/bulk": {
      "post": {
        "description": "Create multiple users in a single transaction-like operation.",
        "operationId": "UsersController_bulkCreate",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkCreateUsersDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Users created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserResponseDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "User email already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "422": {
            "description": "Bulk payload violates business rules",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Bulk create users",
        "tags": [
          "users"
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "description": "Get a single active user by id.",
        "operationId": "UsersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User id",
            "schema": {
              "example": "5b31b752-89d1-4db2-90e8-3f83c2d06a7b",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User returned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get user by id",
        "tags": [
          "users"
        ]
      },
      "patch": {
        "description": "Update user profile, user_email, roles, or user_password.",
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User id",
            "schema": {
              "example": "5b31b752-89d1-4db2-90e8-3f83c2d06a7b",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "User email already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "422": {
            "description": "Business rule violation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Update user",
        "tags": [
          "users"
        ]
      },
      "delete": {
        "description": "Soft delete a user by setting deletedAt without removing data permanently.",
        "operationId": "UsersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User id",
            "schema": {
              "example": "5b31b752-89d1-4db2-90e8-3f83c2d06a7b",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "User soft deleted successfully"
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Soft delete user",
        "tags": [
          "users"
        ]
      }
    }
  },
  "info": {
    "title": "TEST Open API",
    "description": "OpenAPI documentation generated from NestJS source code for Apidog sync.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development"
    },
    {
      "url": "https://staging-api.example.com",
      "description": "Staging"
    },
    {
      "url": "https://api.example.com",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Paste JWT access token without Bearer prefix."
      }
    },
    "schemas": {
...
      "ErrorDetailDto": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string",
            "example": "email",
            "description": "ชื่อ field หรือ path ที่เกิด error"
          },
          "message": {
            "type": "string",
            "example": "email must be an email",
            "description": "รายละเอียด error ของ field นั้น"
          }
        },
        "required": [
          "field",
          "message"
        ]
      },
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false,
            "description": "สถานะสำเร็จของ response"
          },
          "requestId": {
            "type": "string",
            "example": "req_01hzz6qk4vx8b9n2",
            "description": "Request ID สำหรับ trace log และ Sentry"
          },
          "statusCode": {
            "type": "number",
            "example": 400,
            "description": "HTTP status code"
          },
          "error": {
            "type": "string",
            "example": "Bad Request",
            "description": "ประเภทของ error"
          },
          "message": {
            "type": "string",
            "example": "Validation failed",
            "description": "ข้อความ error สำหรับ client"
          },
          "path": {
            "type": "string",
            "example": "/api/v1/users",
            "description": "Path ที่เกิด error"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-05-02T10:30:00.000Z",
            "description": "เวลาที่เกิด error ในรูปแบบ ISO string"
          },
          "details": {
            "description": "รายละเอียด error ราย field",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ErrorDetailDto"
            }
          }
        },
        "required": [
          "success",
          "requestId",
          "statusCode",
          "error",
          "message",
          "path",
          "timestamp"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "user_email": {
            "type": "string",
            "description": "Email ของผู้ใช้ในรูปแบบ field ภายนอก user_email",
            "example": "user@example.com",
            "format": "email"
          },
          "user_password": {
            "type": "string",
            "description": "รหัสผ่านในรูปแบบ field ภายนอก user_password อย่างน้อย 8 ตัวอักษร",
            "example": "password123",
            "minLength": 8
          },
          "fullName": {
            "type": "string",
            "description": "ชื่อ-นามสกุล",
            "example": "สมชาย ใจดี",
            "minLength": 2,
            "maxLength": 160
          },
          "profileImageUrl": {
            "type": "string",
            "description": "URL รูปโปรไฟล์ของผู้ใช้",
            "example": "https://cdn.example.com/users/avatar.png",
            "format": "uri"
          },
          "roles": {
            "description": "รายการ role ของผู้ใช้",
            "example": [
              "user"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "user_email",
          "user_password",
          "fullName"
        ]
      },
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "รหัส user ในระบบ",
            "example": "5b31b752-89d1-4db2-90e8-3f83c2d06a7b",
            "format": "uuid"
          },
          "user_email": {
            "type": "string",
            "description": "Email ของผู้ใช้ในรูปแบบ field ภายนอก user_email",
            "example": "user@example.com",
            "format": "email"
          },
          "fullName": {
            "type": "string",
            "description": "ชื่อ-นามสกุล",
            "example": "สมชาย ใจดี"
          },
          "profileImageUrl": {
            "type": "string",
            "description": "URL รูปโปรไฟล์ของผู้ใช้",
            "example": "https://cdn.example.com/users/avatar.png",
            "nullable": true
          },
          "roles": {
            "description": "รายการ role ของผู้ใช้",
            "example": [
              "user"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "type": "string",
            "description": "เวลาที่สร้าง user",
            "example": "2026-05-03T10:30:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "description": "เวลาที่แก้ไข user ล่าสุด",
            "example": "2026-05-03T10:35:00.000Z"
          }
        },
        "required": [
          "id",
          "user_email",
          "fullName",
          "profileImageUrl",
          "roles",
          "createdAt",
          "updatedAt"
        ]
      },
      "BulkCreateUsersDto": {
        "type": "object",
        "properties": {
          "users": {
            "description": "รายการ users ที่ต้องการสร้าง",
            "minItems": 1,
            "maxItems": 50,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateUserDto"
            }
          }
        },
        "required": [
          "users"
        ]
      },
      "PaginationMetaDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "example": 1,
            "description": "หน้าปัจจุบัน"
          },
          "limit": {
            "type": "number",
            "example": 20,
            "description": "จำนวนรายการต่อหน้า"
          },
          "total": {
            "type": "number",
            "example": 120,
            "description": "จำนวนรายการทั้งหมด"
          },
          "totalPages": {
            "type": "number",
            "example": 6,
            "description": "จำนวนหน้าทั้งหมด"
          }
        },
        "required": [
          "page",
          "limit",
          "total",
          "totalPages"
        ]
      },
      "UserListResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "description": "รายการ users ในหน้าปัจจุบัน",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserResponseDto"
            }
          },
          "meta": {
            "description": "ข้อมูล pagination",
            "allOf": [
              {
                "$ref": "#/components/schemas/PaginationMetaDto"
              }
            ]
          }
        },
        "required": [
          "items",
          "meta"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "user_email": {
            "type": "string",
            "description": "Email ของผู้ใช้ในรูปแบบ field ภายนอก user_email",
            "example": "user@example.com",
            "format": "email"
          },
          "user_password": {
            "type": "string",
            "description": "รหัสผ่านในรูปแบบ field ภายนอก user_password อย่างน้อย 8 ตัวอักษร",
            "example": "password123",
            "minLength": 8
          },
          "fullName": {
            "type": "string",
            "description": "ชื่อ-นามสกุล",
            "example": "สมชาย ใจดี",
            "minLength": 2,
            "maxLength": 160
          },
          "profileImageUrl": {
            "type": "string",
            "description": "URL รูปโปรไฟล์ของผู้ใช้",
            "example": "https://cdn.example.com/users/avatar.png",
            "format": "uri"
          },
          "roles": {
            "description": "รายการ role ของผู้ใช้",
            "example": [
              "user"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    }
  }
}